Move NS communication code into a separate file
